Rename, C API, minimum iOS target, and j make arg
- Renamed libnode.framework to Node.framework
Apple framework naming conventions declare that framework names should start with a capital letter, and the `lib` prefix is reserved for static/dynamic libraries.

- Moved the node::Start method to a node_start C method, allowing Swift compatibility
Swift does not support C++ interop, and importing the libnode framework into an iOS Swift project causes a build error. To fix this, I've wrapped the node::Start method into an equivalent C method, and exposed that in a C header.

- Set minimum iOS target to iOS 9.0
There is no reason to only target the latest version of iOS, as this library does not use new APIs introduced in iOS 11.
I arbitrarily set the target to iOS 9, which is a typical minimum version that you'll find in apps that are backwards compatible.

- Add -j8 to make arguments for iOS target
Adding the `j` flag to make in the ios_framework_prepare script speeds up builds immensely on multi-core machines, by building in parallel, with a maximum of 8 concurrent build runners.